Feature/like

+++ internal/service/memberships/login.go
@@ -29,7 +29,7 @@ func (s *service) Login(request memberships.LoginRequest) (string, error) {
 		return "", errors.New("invalid password")
 	}
 
-	accessToken, err := jwt.CreateToken(int64(userDetail.ID), userDetail.Username, s.cfg.Service.SecretJWT)
+	accessToken, err := jwt.CreateToken(userDetail.ID, userDetail.Username, s.cfg.Service.SecretJWT)
 
 	if err != nil {
 		log.Error().Err(err).Msg("failed to create access token")
